refactor: Improve command handling and socket configuration

+++ Docker/frappe/entrypoint.sh
@@ -28,17 +28,19 @@ if [[ -n "${WORKER_NAME:-}" ]]; then
     SERVICE_NAME="${WORKER_NAME}"
 fi
 
-SOCK_DIR='/fm-sockets'
-SOCK_SERVICE_PATH="$SOCK_DIR/$SERVICE_NAME.sock"
+if [[ -n "${WORKER_NAME:-}" || -n "${SERVER_NAME:-}" ]]; then
+    SOCK_DIR='/fm-sockets'
+    SOCK_SERVICE_PATH="$SOCK_DIR/$SERVICE_NAME.sock"
 
-echo "Setting supervisord sock directory to $SOCK_SERVICE_PATH"
+    echo "Setting supervisord sock directory to $SOCK_SERVICE_PATH"
 
-mkdir -p /opt/user/conf.d $SOCK_DIR
-chown "$USERID:$USERGROUP" $SOCK_DIR /opt/user /opt/user/conf.d
-rm -rf "$SOCK_SERVICE_PATH"
+    mkdir -p /opt/user/conf.d $SOCK_DIR
+    chown "$USERID:$USERGROUP" $SOCK_DIR /opt/user /opt/user/conf.d
+    rm -rf "$SOCK_SERVICE_PATH"
 
-sed -i "s/\opt\/user\/supervisor\.sock/fm-sockets\/${SERVICE_NAME}\.sock/g" /opt/user/supervisord.conf
-echo "supervisord configured $?"
+    sed -i "s/\opt\/user\/supervisor\.sock/fm-sockets\/${SERVICE_NAME}\.sock/g" /opt/user/supervisord.conf
+    echo "supervisord configured $?"
+fi
 
 if [ "$#" -gt 0 ]; then
     gosu "$USERID":"$USERGROUP" "/scripts/$@" &
